+++ src/encoder_constants.c
@@ -1,3 +1,3 @@
 #include "constants/encoder_constants.h"
 
-const int B = CLIP(OUTPUT_WORD_SIZE, 1, 8);
+const uint8_t B = (uint8_t) (CLIP(OUTPUT_WORD_SIZE, 1, 8));
